Mirror API accepts an additional identifier parameter.
Since multiple clients can use the mirror API at the same time, a
client can now send an identifier string with their API call. This
token is included in the API response and allows clients to filter out
responses for calls they didn't make.

The API input string now follows the format
<identifier>\n<action>\n<pluginName>\n<values>

And the response follows the format
<identifier>:<action>:<status>:<pluginName>:<payload>

If no identifier is required, it can be left blank for the API call:
\n<action>\n<pluginName>\n<values>

Which means its not backwards compatible.